+++ test/binary_test.exs
@@ -10,7 +10,6 @@ defmodule BinaryTest do
   setup do
     opts = [
       hostname: "sqlserver.local",
-      port: 4000,
       username: "test_user",
       password: "passw0rd!",
       database: "test_db"