+++ test/query_test.exs
@@ -9,7 +9,6 @@ defmodule QueryTest do
   setup do
     opts = [
       hostname: "sqlserver.local",
-      port: 4000,
       username: "test_user",
       password: "passw0rd!",
       database: "test_db"