+++ test/test_helper.exs
@@ -1,3 +1,4 @@
+System.at_exit fn _ -> Logger.flush end
 ExUnit.start()