+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   * Fixed: Packets sent to the server which exceed the negotiated packet size would cause the connection to close  
 * Enhancements
   * Added support for decoding Time(n) and DateTime2
+  * Added support for SQL Named Instances, pass instance: "instance_name" in connection options
 
 # v0.2.0
 * Enhancements